add filter by category

+++ app/Http/Controllers/BooksController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 use App\Models\books;
+use App\Models\categories;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class BooksController extends Controller
 {
@@ -39,27 +41,42 @@ class BooksController extends Controller
     }
     public function updateBook(request $req , $id){
         $data = books::find($id);
-        $imgName=$req->file('image')->getClientOriginalName();
-        $req->file('image')->move(public_path('img'), $imgName);
-        $data->update([
-            'title'=>$req->title,
-            'description'=>$req->description,
-            'price'=>$req->price,
-            'category_id'=>$req->category_id,
-            'user_id'=>$req->user()->id,
-            'image'=>$imgName,
-            'isbn'=>$req->isbn,
-            'auteur'=>$req->auteur
-        ]);
-        return response()->json(['message' => "the book has been updated successfully", 'book'=>$data]);
+        $user = Auth::user();
+        if($user->can('edit every book') || $user->id == $data->user_id){
+            $imgName=$req->file('image')->getClientOriginalName();
+            $req->file('image')->move(public_path('img'), $imgName);
+            $data->update([
+                'title'=>$req->title,
+                'description'=>$req->description,
+                'price'=>$req->price,
+                'category_id'=>$req->category_id,
+                'image'=>$imgName,
+                'isbn'=>$req->isbn,
+                'auteur'=>$req->auteur
+            ]);
+            return response()->json(['message' => "the book has been updated successfully", 'book'=>$data]);
+        }
+            return response()->json(['message' => "you don't have permission to update this book"]);
+
     }
     public function deleteBook($id){
         $data = books::find($id);
-        if ($data) {
-            $data->delete();
-            return response()->json(['message' => "the book has been deleted successfully", 'book'=>$data]);
-        } else {
-            return response()->json(['message' => "the book you want to delete doesn't exist!"]);
+        $user = Auth::user();
+        if($user->can('edit every book') || $user->id == $data->user_id){
+            if ($data) {
+                $data->delete();
+                return response()->json(['message' => "the book has been deleted successfully", 'book'=>$data]);
+            } else {
+                return response()->json(['message' => "the book you want to delete doesn't exist!"]);
+            }
         }
+        return response()->json(['message' => "you don't have permission to delete this book"]);
+
+    }
+    public function filter($category)
+    {
+        $category = categories::where('category', $category)->firstOrFail();
+        $books = books::where('category_id', $category->id)->get();
+        return response()->json($books);
     }
 }
